// src/lib/api/types/adminCategoryDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface AdminCategoryDTO {
  /** @nullable */
  children?: AdminCategoryDTO[] | null;
  description?: string;
  icon?: string;
  id: string;
  name: string;
  parentId?: string;
  slug: string;
  sortOrder: number;
}